Practica Integradora 3 (#10)
* practica integradora entrega 3

+++ src/services/ProductService.ts
@@ -1,11 +1,12 @@
 import mongoose, { PaginateResult } from 'mongoose'
-import DaoFactory from '../dao/DaoFactory'
-import ProductRepository from '../dao/mongo/repository/ProductRepository'
+import DaoFactory from '../persistence/DaoFactory'
+import ProductRepository from '../persistence/mongo/repository/ProductRepository'
 import { Product } from '../entities/IProduct'
-import ProductResponseDTO from '../dao/mongo/dtos/product/Product.response'
+import { UserRoles } from '../entities/IUser'
 import { sendSocketMessage } from '../socket/SocketClient'
 import createError from 'http-errors'
 import { logger } from '../utils/Logger'
+
 export default class ProductService {
     private static instance: ProductService | null = null
     private productManager
@@ -22,7 +23,7 @@ export default class ProductService {
         }
         return ProductService.instance
     }
-    
+
     async addProduct(data: any): Promise<Product> {
         try {
             const savedProduct = await this.productManager.addProduct(data)
@@ -42,7 +43,7 @@ export default class ProductService {
         }
     }
 
-    async getProducts(pipeline?:any, options?:any): Promise<PaginateResult<Product> | any> {
+    async getProducts(pipeline?: any, options?: any): Promise<PaginateResult<Product> | any> {
         return await this.productRepository.getProducts(pipeline, options)
     }
 
@@ -56,57 +57,79 @@ export default class ProductService {
         }
     }
 
-    async deleteProductById(id: any): Promise<void> {
+    async deleteProductById(id: any, userRole: string): Promise<void> {
+        const product = await this.getProductById(id)
+        this.validateUserRole(userRole, product)
         const deleted = await this.productManager.deleteProductById(id)
-        if (deleted === 1) {
+        
+        if (deleted == 1) {
             sendSocketMessage("productDeleted", id)
+            logger.debug(`product with id: ${id} deleted`)
         } else {
             throw new createError.NotFound(`Product ${id} not found`)
         }
     }
 
-    async deleteAll(): Promise<void> {
+    async deleteAll(userRole: string): Promise<void> {
+        if (userRole === UserRoles.PREMIUM) {
+            throw new createError
+                .NotFound(`User not authorized to perform this action`)
+        }
         await this.productManager.deleteAll()
     }
 
-    async updateProductById(prodId: string, updatedFields: Partial<Product>): Promise<Product> {
+    async updateProductById(prodId: string, updatedFields: Partial<Product>,
+        userRole: string): Promise<Product | void> {
         try {
-            let product = await this.productManager.updateProductById(prodId, updatedFields)
-            sendSocketMessage("productSaved", product)
-    
-            return product
-        } catch (error:any) {
+            const product = await this.getProductById(prodId)
+            this.validateUserRole(userRole, product)
+
+            let udpatedProduct = await this.productManager
+                .updateProductById(prodId, updatedFields)
+            sendSocketMessage("productSaved", udpatedProduct)
+
+            return udpatedProduct
+        } catch (error: any) {
             throw new createError.NotFound(error.message)
+        }
+    }
+
+    private validateUserRole(userRole: string, product: Product): void {
+        const isUserAuthorized = (userRole === UserRoles.ADMIN ||
+            (userRole === UserRoles.PREMIUM && product.owner === UserRoles.PREMIUM))
+        if (!isUserAuthorized) {
+            throw new createError
+                .NotFound(`User not authorized to perform this action`)
         }
     }
 
     buildProductQueryPipeline(category: any, status: any, sort: any): any[] {
         const parsedStatus = status === 'true'
         const pipeline: any[] = []
-      
+
         if (category) {
-          pipeline.push({
-            $match: {
-              category: category,
-            },
-          })
+            pipeline.push({
+                $match: {
+                    category: category,
+                },
+            })
         }
-      
+
         if (status !== undefined) {
-          pipeline.push({
-            $match: {
-              status: parsedStatus,
-            },
-          })
+            pipeline.push({
+                $match: {
+                    status: parsedStatus,
+                },
+            })
         }
-      
+
         if (sort) {
-          const sortOrder = parseInt(sort)
-          pipeline.push({
-            $sort: {
-              price: sortOrder,
-            },
-          })
+            const sortOrder = parseInt(sort)
+            pipeline.push({
+                $sort: {
+                    price: sortOrder,
+                },
+            })
         }
         logger.debug(`pipeline: ${JSON.stringify(pipeline)}`)
         return pipeline
